Refactor cluster_info module to reduce complexity of the code
